multi-session: add multiple session capability
This CR nests multiple logins under one key.  This is a nice property
for operators because it keeps an individual's login session in one
place instead of having to build a lookup index elsewhere mapping
user->active sessions.

It uses lazy deletion and expiration checks.  There's a global
expiration which exists for memcache or redis auto-expiration.

There exists an inherent race if multiple sessions are being set at
once.  This race results in the last Set succeeding.  In practice, an
individual user shouldn't realistically be logging into two or more
devices at the same time.