context: add basic support for context
This doesn't add full support.  Doing so requires cooperation from the
libraries.  However, this does solve the case where a request is
cancelled, we don't block and return control to the caller.  The
consequence is that we're leaking a goroutine for the remainder of the
open request to memcache/redis (which shouldn't be long)

Both libraries expose the ability to set timeouts on connections.  It
might be worthwhile to see how to integrate them more seamlessly:

https://godoc.org/github.com/bradfitz/gomemcache/memcache#ConnectTimeoutError
https://godoc.org/github.com/gomodule/redigo/redis#DoWithTimeout