Remove error message when using Insecure option
I understand the purpose of the Secure attribute on cookies, but I think the jeff library is making too many assumptions about requiring them. The error message is inappropriate in several situations:

- During development, the log message creates unnecessary output on every reload of the server code.
- There are instances where an implementer might have a legitimate need to exclude the Secure cookie attribute (e.g., a trusted network, an HTTP connection that's routed through a secure SSH tunnel).

To remove the Secure attribute, jeff already requires the client to specify the Insecure option. If the user is explicitly choosing that option, the library should accept that choice rather than generating error messages about it.